+++ packaging/install-files.sh
@@ -47,6 +47,7 @@ install -d -m 0755 \
     "$DOC_DIR"
 
 for file in "$PROJECT_DIR"/configurator/*.py; do
+    [ -e "$file" ] || continue
     install -m 0644 "$file" "$APP_DIR/configurator/"
 done
 for file in \
@@ -57,6 +58,7 @@ do
     install -m 0644 "$PROJECT_DIR/$file" "$APP_DIR/$file"
 done
 for file in "$PROJECT_DIR"/script/*.sh; do
+    [ -e "$file" ] || continue
     install -m 0755 "$file" "$APP_DIR/script/"
 done
